split startup into several files
